requirements.txt
openai
python-dotenv
chromadb
fastapi
uvicorn
httpx
